Add ugly layout manager as a temporary solution for solving the problem on having two trees communicating with each other
